+++ data/trx/ship/games/tr3/scripts/rapids.lua
@@ -1,0 +1,5 @@
+trx.events.before_item_setup(function(level)
+  -- Setup shoals
+  trx.items[16].properties.range = { x = 6, y = 5, z = 10 }
+  trx.items[52].properties.range = { x = 18, y = 8, z = 18 }
+end)